Don't give warning that readonly variable is forced to be readonly mysql-test-run run now fails if we have [Warning] and [ERROR] as tags in .err file Fixed wrong reference to the mysql manual Fixed wrong prototype that caused some tests to fail on 64 bit platforms
mysql-test/lib/mtr_report.pl:
  Test run now fails if we have [Warning] and [ERROR] as tags in .err file
  Added list of all common 'not fatal' errors to ignore error list
mysql-test/mysql-test-run-shell.sh:
  Fixed some wrong startup options
mysql-test/t/disabled.def:
  Disable instance manager tests because they generate warnings (and probably don't read the option files correctly)
sql/ha_ndbcluster_binlog.cc:
  Ensure we log all binglog errors with the "NDB Binlog" tag
sql/slave.cc:
  Make errors uniform
sql/sql_plugin.cc:
  Don't give warning that readonly variable is forced to be readonly
sql/stacktrace.c:
  Corrected manual reference
storage/blackhole/ha_blackhole.cc:
  Fixed wrong prototype that caused test to fail on 64 bit platforms
storage/example/ha_example.cc:
  Fixed wrong prototype that caused test to fail on 64 bit platforms

+++ storage/blackhole/ha_blackhole.cc
@@ -287,8 +287,8 @@ static void blackhole_free_key(st_blackhole_share *share)
   my_free((uchar*) share, MYF(0));
 }
 
-static uchar* blackhole_get_key(st_blackhole_share *share, uint *length,
-                               my_bool not_used __attribute__((unused)))
+static uchar* blackhole_get_key(st_blackhole_share *share, size_t *length,
+                                my_bool not_used __attribute__((unused)))
 {
   *length= share->table_name_length;
   return (uchar*) share->table_name;
